requirements.txt
# ─── CryptoTrader Pro — Python dependencies ───────────────────────────────────
# Install: pip install -r requirements.txt

# ── Core async & HTTP ──────────────────────────────────────────────────────────
aiohttp==3.9.5
websockets==12.0
httpx==0.27.0

# ── FastAPI backend ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
slowapi==0.1.9

# ── Auth & security ────────────────────────────────────────────────────────────
PyJWT==2.8.0
bcrypt==4.1.3
cryptography==42.0.5
pyotp==2.9.0

# ── Database ───────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.29
alembic==1.13.1
# For PostgreSQL support (uncomment if needed):
# psycopg2-binary==2.9.9

# ── Data & ML ──────────────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
joblib==1.4.0

# ── Dashboard ──────────────────────────────────────────────────────────────────
streamlit==1.34.0
plotly==5.22.0

# ── Notifications ──────────────────────────────────────────────────────────────
# Telegram via aiohttp (built-in)
# Discord via aiohttp webhook (built-in)
# Email via smtplib (built-in)

# ── Cache ──────────────────────────────────────────────────────────────────────
redis==5.0.3       # optional — only used when USE_REDIS=true

# ── Config ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Dev / testing ──────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0     # for FastAPI TestClient
